_Edit, _List を path.svelte に統合
