+++ src/routes/[...path].svelte
@@ -36,7 +36,6 @@
     return {
       props: {
         path,
-        paths,
         mode,
         content,
         id,
@@ -49,17 +48,98 @@
   import { goto } from "$app/navigation";
   import { Meta } from "svelte-head";
 
-  import List from "./_List.svelte";
-  import Edit from "./_Edit.svelte";
-  import {Header} from "svelte-admin-components";
+  import { ContentList, ContentForm, Header } from "svelte-admin-components";
+  import { indicator } from "svelte-modal-manager";
 
   export let path;
-  export let paths;
   export let mode;
   export let content;
   export let id;
 
-  content = admin.actions.pathToContent(path);
+  let actions = admin.actions;
+  let item;
+  let form;
+
+  // ライブリロード
+  content = actions.pathToContent(path);
+
+  // 選択
+  let onSelect = (e, item) => {
+    // 遷移する
+    let url = `/${path}/${e.detail.item.id}`;
+    if (e.detail.originalEvent.metaKey) {
+      // 新しいタブで開く
+      window.open(url);
+    }
+    else {
+      goto(url);
+    }
+  };
+
+  // item 取得
+  let fetchItem = async () => {
+    if (id === 'new') return ;
+
+    ({item} = await actions.api.get({
+      path,
+    }));
+  };
+
+  // 新規作成 / 更新
+  let onSubmit = async (e) => {
+    let item = e.detail.value;
+
+    let i = indicator();
+
+    try {
+      if (id !== 'new') {
+        // 更新
+        let res = await actions.api.put({
+          path,
+          data: item,
+        });
+        console.log('saved', res);
+      }
+      else {
+        // 新規作成
+        let res = await actions.api.post({
+          path: path.replace(/\/new$/, ''),
+          data: item,
+        });
+
+        console.log('created', res);
+        // 単体編集ページに遷移
+        let to = path.replace(/new$/, res.item.id);
+        goto(`/${to}`, {
+          replaceState: true,
+        });
+      }
+    }
+    catch(e) {
+      alert(e);
+    }
+    finally {
+      setTimeout(() => {
+        i.close();
+      }, 512);
+    }
+  };
+
+  // 削除
+  let onDelete = async () => {
+    if (!confirm('really?')) {
+      return ;
+    }
+
+    // 削除
+    let res = await actions.api.del({
+      path,
+    });
+    console.log('deleted', res);
+
+    // 前のページに戻る
+    history.back();
+  };
 
   let getHeaderLabel = (path) => {
     return id ? `${content.label} / ${id}` : content.label;
@@ -85,7 +165,7 @@
           label: 'CREATE',
           kind: 'primary',
           onclick: () => {
-            alert();
+            form.submit();
           }
         },
       ];
@@ -93,23 +173,30 @@
     else {
       // 編集
       return [
+        // 削除
         {
           label: 'DELETE',
           kind: 'danger',
-          onclick: () => {
-            alert();
+          onclick: (e) => {
+            onDelete(e);
           }
         },
+        // 保存
         {
           label: 'SAVE',
           kind: 'primary',
           onclick: () => {
-            alert();
+            form.submit();
           }
         },
       ];
     }
   };
+
+  $: {
+    path;
+    fetchItem();
+  }
 </script>
 
 <Meta />
@@ -118,10 +205,11 @@
   main.s-full.overflow-scroll
     Header.p16.sticky.t0.box-shadow.bg-white.relative.z100(label='{getHeaderLabel(path)}', actions='{getHeaderActions(path)}')
 
-    +if('mode === "list"')
-      List(content='{content}', path='{path}', actions='{admin.actions}', id='{id}')
-    +if('mode === "edit"')
-      Edit(content='{content}', path='{path}', actions='{admin.actions}', id='{id}')
-    +if('mode === "schema"')
-      div TODO: schema 編集
+    div.p16
+      +if('mode === "list"')
+        div.bg-white.box-shadow.rounded-4.mb16
+          ContentList(path='{path}', content='{content}', actions='{actions}', limit='{16}', on:select='{onSelect}')
+        //- List(content='{content}', path='{path}', actions='{actions}', id='{id}')
+      +if('mode === "edit"')
+        ContentForm(bind:this='{form}', path='{path}', value='{item}', sections='{content.sections}', actions='{admin.actions}', on:submit='{onSubmit}')
 </template>
